app/requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
cffi==2.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.23